[issue_304] usage-state caching
Updated error message on username-claim renewal failure.

Updated  SpaceCheckingKeyFilter to store/load (username -> usage) state
to/from a local file, falling back to re-calculating it from scratch on
startup.

Note, for the shutdown-hook to be triggered either a kill signal with the
default level (or sending a ctrl-c) should be used to kill the process.